+++ libraries/objects/mod.ts
@@ -1,11 +1,13 @@
 export { hasOwnProperty } from './src/hasOwnProperty.ts';
-export type { PlainObject } from './src/checkIsPlainObject.ts';
-export { checkIsPlainObject, isPlainObject } from './src/checkIsPlainObject.ts';
+export { isNonNullObject } from './src/isNonNullObject.ts';
+export type { PlainObject } from './src/isPlainObject.ts';
+export { checkIsPlainObject, isPlainObject } from './src/isPlainObject.ts';
+export { isScalar } from './src/isScalar.ts';
 export { mapToObject } from './src/mapToObject.ts';
 export type { NonNullishProperties } from './src/omitNullish.ts';
 export { omitNullish } from './src/omitNullish.ts';
 export type { DefinedProperties } from './src/omitUndefined.ts';
 export { omitUndefined } from './src/omitUndefined.ts';
-export { preciseObjectTypeof } from './src/preciseObjectTypeof.ts';
+export { preciseObjectTypeof, preciseTypeOf } from './src/preciseTypeOf.ts';
 export { sortKeys } from './src/sortKeys.ts';
 export { sortObjectKeys } from './src/sortObjectKeys.ts';